Enable remote IO for polars benchmarks
This enables the benchmarking of polars from a remote storage source,
like S3. A new `StorageOptions` class is added to `settings.py` to
configure things like the endpoint URL, region, and credentials.

I needed to update some of the path handling to work with either (local)
paths represented as pathlib.Path objects or remote paths represented as
strings.